requirements.txt
aiogram==3.8.0
fastapi==0.111.0
python-dotenv==1.0.1
mongoengine==0.28.2